# openapi.json
{"openapi":"3.1.0","info":{"title":"Recipe API","version":"1.0.0"},"paths":{"/api/recipes/":{"post":{"tags":["recipes"],"summary":"Create Recipe","description":"Create a new recipe.","operationId":"create_recipe_api_v1_recipes__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecipeCreate"}}}},"responses":{"200":{"description":"Recipe created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Recipe"}}}},"422":{"description":"Invalid recipe data"}}},"get":{"tags":["recipes"],"summary":"List recipes","description":"List recipes with optional filters \n        Query parameters: vegetarian, servings, include_ingredients, exclude_ingredients, instructions_search","operationId":"list_recipes_api_v1_recipes__get","parameters":[{"name":"vegetarian","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"description":"Filter by vegetarian","title":"Vegetarian"},"description":"Filter by vegetarian"},{"name":"servings","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"description":"Filter by number of servings","title":"Servings"},"description":"Filter by number of servings"},{"name":"include_ingredients","in":"query","required":false,"schema":{"type":"array","items":{"type":"string"},"description":"Recipes that include any of these ingredients","default":[],"title":"Include Ingredients"},"description":"Recipes that include any of these ingredients"},{"name":"exclude_ingredients","in":"query","required":false,"schema":{"type":"array","items":{"type":"string"},"description":"Recipes that exclude any of these ingredients","default":[],"title":"Exclude Ingredients"},"description":"Recipes that exclude any of these ingredients"},{"name":"instructions_search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Full-text search in instructions","title":"Instructions Search"},"description":"Full-text search in instructions"}],"responses":{"200":{"description":"List of recipes","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Recipe"},"title":"Response List Recipes Api V1 Recipes  Get"}}}},"422":{"description":"Invalid filter data"}}}},"/api/recipes/{recipe_id}":{"get":{"tags":["recipes"],"summary":"Get Recipe","description":"Get a recipe by ID","operationId":"get_recipe_api_v1_recipes__recipe_id__get","parameters":[{"name":"recipe_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Recipe Id"}}],"responses":{"200":{"description":"Recipe found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Recipe"}}}},"404":{"description":"Recipe not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["recipes"],"summary":"Update Recipe","description":"Update an existing recipe by ID","operationId":"update_recipe_api_v1_recipes__recipe_id__put","parameters":[{"name":"recipe_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Recipe Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecipeUpdate"}}}},"responses":{"200":{"description":"Recipe updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Recipe"}}}},"404":{"description":"Recipe not found"},"422":{"description":"Invalid recipe data"}}},"delete":{"tags":["recipes"],"summary":"Delete Recipe","description":"Delete a recipe by ID. Returns 200 on success. Returns 404 if the recipe does not exist.","operationId":"delete_recipe_api_v1_recipes__recipe_id__delete","parameters":[{"name":"recipe_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Recipe Id"}}],"responses":{"200":{"description":"Recipe deleted","content":{"application/json":{"schema":{}}}},"404":{"description":"Recipe not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Ingredient":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name"},"unit":{"type":"string","title":"Unit"},"quantity":{"type":"integer","title":"Quantity"}},"type":"object","required":["id","name","unit","quantity"],"title":"Ingredient"},"IngredientCreate":{"properties":{"name":{"type":"string","title":"Name"},"unit":{"type":"string","title":"Unit"},"quantity":{"type":"integer","title":"Quantity"}},"type":"object","required":["name","unit","quantity"],"title":"IngredientCreate"},"Recipe":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"vegetarian":{"type":"boolean","title":"Vegetarian"},"servings":{"type":"integer","title":"Servings"},"instructions":{"type":"string","title":"Instructions"},"ingredients":{"items":{"$ref":"#/components/schemas/Ingredient"},"type":"array","title":"Ingredients"}},"type":"object","required":["id","vegetarian","servings","instructions"],"title":"Recipe"},"RecipeCreate":{"properties":{"vegetarian":{"type":"boolean","title":"Vegetarian"},"servings":{"type":"integer","title":"Servings"},"instructions":{"type":"string","title":"Instructions"},"ingredients":{"items":{"$ref":"#/components/schemas/IngredientCreate"},"type":"array","title":"Ingredients"}},"type":"object","required":["vegetarian","servings","instructions"],"title":"RecipeCreate"},"RecipeUpdate":{"properties":{"vegetarian":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Vegetarian"},"servings":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Servings"},"instructions":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Instructions"}},"type":"object","title":"RecipeUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}